Cambio de visualizador a componente más arreglo completo de errores en run build
