+++ app/chatbot/page.tsx
@@ -3,7 +3,6 @@
 'use client';
 import { useSession, signIn } from 'next-auth/react';
 import { useState } from 'react';
-import { useRouter } from 'next/navigation';
 
 const ChatbotPage = () => {
   const { data: session, status } = useSession();
