+++ app/components/Acordion.tsx
@@ -4,9 +4,9 @@ const Acordion = ({nombre,informacion}:{nombre:string, informacion:string}) => {
             <div className="py-5">
                 <details className="group">
                     <summary className="flex justify-between items-center font-medium cursor-pointer list-none">
-                        <span>{nombre}</span>
+                        <span className="text-neutral-50">{nombre}</span>
                         <span className="transition group-open:rotate-180">
-                <svg fill="none" height="24" shapeRendering="geometricPrecision" stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5" viewBox="0 0 24 24" width="24"><path d="M6 9l6 6 6-6"></path>
+                <svg fill="none" height="24" shapeRendering="geometricPrecision" stroke="white" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5" viewBox="0 0 24 24" width="24"><path d="M6 9l6 6 6-6"></path>
                 </svg>
                 </span>
                     </summary>
