+++ app/components/Modal.tsx
@@ -1,5 +1,3 @@
-import { useState } from "react";
-
 interface ModalProps {
   isOpen: boolean;
   onClose: () => void;
