+++ next.config.mjs
@@ -1,5 +1,15 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
-
+const nextConfig = {
+    images:{ remotePatterns: [
+        {
+            protocol: "https",
+            hostname: "www.freeiconspng.com"
+        },
+        {
+            protocol: "https",
+            hostname: "images.unsplash.com"
+        }
+    ]}
+};
 
 export default nextConfig;
